Fixes
Added current/max for Derived Stats. (Maby someone wants to add the modification system to this too).

Change the Values from Value to Current and Max to be uniform with statistics.

Added Image to loot sheet. Localized the Proffession sheet and added editor boxes.

Added some localization and misc fixes